refactor for, printf, template to homogeneous spelling

+++ include/radial_r2grid.hxx
@@ -7,18 +7,18 @@
 
 namespace radial_r2grid {
 
-  template <typename real_t=float>
+  template <typename real_t=double>
   std::vector<real_t> r_axis(int const nr2, float const ar2=1) {
-      float const ar2inv = 1.0/ar2;
+      double const ar2inv = 1.0/ar2;
       std::vector<real_t> r(nr2);
-      for(int ir2 = 0; ir2 < nr2; ++ir2) {
+      for (int ir2 = 0; ir2 < nr2; ++ir2) {
           r[ir2] = std::sqrt(ir2*ar2inv); // r^2-grid formula
           // r^2-grid inversion:      i0 = int(r2*ar2)
           //    --> interpolate between i0 and i0 + 1
       } // ir2
       return r;
   } // r_axis
- 
+
   inline status_t all_tests(int const echo=0) { return STATUS_TEST_NOT_INCLUDED; }
-  
+
 } // namespace radial_r2grid
